requirements.txt
pandas
scikit-learn
numpy
joblib
requests
python-dotenv
kiteconnect
